+++ delayed_image/__init__.py
@@ -1,9 +1,5 @@
 r"""
-A rewrite of the delayed operations
-
-Note:
-    The classes in this submodule will have their names changed when the old
-    POC delayed operations are deprecated.
+The delayed image module.
 
 TODO:
     The optimize logic could likley be better expressed as some sort of