+++ CHANGELOG.md
@@ -5,7 +5,15 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.3 - Unreleased
+## Version 0.2.4 - Unreleased
+
+### Changed
+
+* modified the signature of `DelayedLoad.demo` to better match `__init__` and
+  `grab_image_test_fpath`.
+
+
+## Version 0.2.3 - Released 2022-11-07
 
 ### Fixed
 * Issue in SensorChan spec where duplicate sensors were present in concise codes